requirements.txt
aiohttp
aiopg
PyYAML
https://github.com/Rapptz/discord.py/archive/rewrite.zip
sqlalchemy
